# app/Models/Product.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Product extends Model {
    public $timestamps = false;
    protected $fillable = ['name', 'price', 'inventory'];
    protected $casts = [
        'price' => 'float',
        'inventory' => 'integer',
    ];
}